Add files via upload
S1_file.pfd is the pdf version of the R markdown used to run the GAM and GLM for the Mada paper. 
S2_file.pfd is the pdf version of the R markdown used to run the MFA for the Mada paper. 
Catches.txt and SM_MFA.txt are the data files respectively associated with S1 and S2_file.pdf